Document chemistry and integrator code
